Change slider via "set" button produces error #1074

+++ src/main/java/li/klass/fhem/adapter/devices/genericui/StateChangingSeekBarFullWidth.java
@@ -51,7 +51,7 @@ public class StateChangingSeekBarFullWidth extends SeekBarActionRowFullWidthAndB
     }
 
     @Override
-    public void onButtonSetValue(XmlListDevice device, int value) {
+    public void onButtonSetValue(XmlListDevice device, float value) {
         onStopTrackingTouch(context, device, value);
     }
